+++ VerbTrainerAuth/Application/Exceptions/UserAlreadyExistsException.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-namespace VerbTrainerAuth.Application.Exceptions
+namespace VerbTrainerUser.Application.Exceptions
 {
 	public class UserAlreadyExistsException : Exception
     {
