+++ VerbTrainerAuth/Domain/Interfaces/IAsyncAccessTokenRepository.cs
@@ -1,8 +1,8 @@
 ﻿using System;
-using VerbTrainerAuth.Domain.Base;
-using Models = VerbTrainerAuth.Infrastructure.Data.Models;
+using VerbTrainerUser.Domain.Base;
+using Models = VerbTrainerUser.Infrastructure.Data.Models;
 
-namespace VerbTrainerAuth.Domain.Interfaces
+namespace VerbTrainerUser.Domain.Interfaces
 {
 	public interface IAsyncAccessTokenRepository : IAsyncRepository<Models.RevokedAccessToken>
     {
