+++ VerbTrainerAuth/Infrastructure/Data/Models/RevokedRefreshToken.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace VerbTrainerAuth.Infrastructure.Data.Models
+namespace VerbTrainerUser.Infrastructure.Data.Models
 {
 	public class RevokedRefreshToken : BaseAuthModel
 	{
