+++ VerbTrainerAuth/Program.cs
@@ -1,18 +1,18 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.IdentityModel.Tokens;
-using VerbTrainerAuth.Infrastructure.Data;
-using VerbTrainerAuth.Infrastructure.Messaging.Configuration;
-using VerbTrainerAuth.Infrastructure.Messaging.Producer;
-using VerbTrainerAuth.Domain.Interfaces;
-using VerbTrainerAuth.Infrastructure.Data.Repositories;
-using VerbTrainerAuth.Application.Services.Mapping;
-using VerbTrainerAuth.Application.Services.JWT;
-using VerbTrainerAuth.Application.Services.User;
-using VerbTrainerAuth.Application.UserLogin;
-using VerbTrainerAuth.Application.RefreshUserAccess;
-using VerbTrainerAuth.Application.RegisterUser;
-using VerbTrainerAuth.Application.ResetPassword;
+using VerbTrainerUser.Infrastructure.Data;
+using VerbTrainerUser.Infrastructure.Messaging.Configuration;
+using VerbTrainerUser.Infrastructure.Messaging.Producer;
+using VerbTrainerUser.Domain.Interfaces;
+using VerbTrainerUser.Infrastructure.Data.Repositories;
+using VerbTrainerUser.Application.Services.Mapping;
+using VerbTrainerUser.Application.Services.JWT;
+using VerbTrainerUser.Application.Services.User;
+using VerbTrainerUser.Application.UserLogin;
+using VerbTrainerUser.Application.RefreshUserAccess;
+using VerbTrainerUser.Application.UserRegister;
+using VerbTrainerUser.Application.ResetPassword;
 using Microsoft.IdentityModel.Logging;
 using VerbTrainerEmail.Infrastructure.Repositories;
