+++ VerbTrainerAuth/VerbTrainerUser.csproj
@@ -40,7 +40,6 @@
     <None Remove="Domain\Exceptions\" />
     <None Remove="Application\UserLogin\" />
     <None Remove="Application\RefreshUserAccess\" />
-    <None Remove="Application\RegisterUser\" />
     <None Remove="Application\Exceptions\" />
     <None Remove="Application\ResetPassword\" />
     <None Remove="Application\Services\" />
@@ -50,6 +49,7 @@
     <None Remove="Application\ChangePassword\" />
     <None Remove="Application\DeleteUser\" />
     <None Remove="Application\SetUserName\" />
+    <None Remove="Application\UserRegister\" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Infrastructure\" />
@@ -66,7 +66,6 @@
     <Folder Include="Domain\Exceptions\" />
     <Folder Include="Application\UserLogin\" />
     <Folder Include="Application\RefreshUserAccess\" />
-    <Folder Include="Application\RegisterUser\" />
     <Folder Include="Application\Exceptions\" />
     <Folder Include="Application\ResetPassword\" />
     <Folder Include="Application\Services\" />
@@ -76,6 +75,7 @@
     <Folder Include="Application\ChangePassword\" />
     <Folder Include="Application\DeleteUser\" />
     <Folder Include="Application\SetUserName\" />
+    <Folder Include="Application\UserRegister\" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\VerbTrainerMessaging\VerbTrainerMessaging.csproj">
